backend/requirements.txt
# ==============================================
# University Attendance System Dependencies
# ==============================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
email-validator==2.1.0

# Face Recognition & Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
scikit-learn==1.3.2
numpy==1.24.3
Pillow==10.1.0

# Machine Learning (Optional - for advanced features)
tensorflow==2.15.0

# Data Processing
pandas==2.1.3
scipy==1.11.4

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Templates & Email
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Caching
redis==5.0.1

# File Processing
python-magic==0.4.27

# Logging
structlog==23.2.0

# Development Dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Deployment
docker==6.1.3
